chore: rename to (ts|tsx)
