--- src/components/EvidenceBuilder/EvidenceBuilder.stories.jsx
@@ -1,10 +1,0 @@
-import { EvidenceBuilder } from "./index";
-
-export default {
-  title: "EvidenceBuilder",
-  component: EvidenceBuilder
-};
-
-const Template = (arg) => <EvidenceBuilder {...arg} />;
-
-export const Variant1 = Template.bind({});
