+++ src/components/Layout/index.ts
@@ -1,6 +1,6 @@
 import { SearchLayout } from "./search";
 import { MinimalLayout } from "./minimal";
-import { SidebarLayout } from "./sidebar";
+import { SidebarLayout } from "./SidebarLayout";
 import { ConfigLayout } from "./config";
 
 export { SearchLayout, SidebarLayout, MinimalLayout, ConfigLayout };
